4. Формы. Загрузка файла на сервер. [#504855] - Files added to Vacancies

+++ src/main/java/ru/job4j/dreamjob/service/SimpleVacancyService.java
@@ -1,6 +1,7 @@
 package ru.job4j.dreamjob.service;
 
 import org.springframework.stereotype.Service;
+import ru.job4j.dreamjob.dto.FileDto;
 import ru.job4j.dreamjob.model.Vacancy;
 import ru.job4j.dreamjob.repository.VacancyRepository;
 
@@ -12,23 +13,48 @@ public class SimpleVacancyService implements VacancyService {
 
     private final VacancyRepository vacancyRepository;
 
-    public SimpleVacancyService(VacancyRepository vacancyRepository) {
-        this.vacancyRepository = vacancyRepository;
+    private final FileService fileService;
+
+    public SimpleVacancyService(VacancyRepository sql2ovacancyRepository,
+                                FileService fileService) {
+        this.vacancyRepository = sql2ovacancyRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Vacancy save(Vacancy vacancy) {
+    public Vacancy save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
         return vacancyRepository.save(vacancy);
     }
 
-    @Override
-    public boolean deleteById(int id) {
-        return vacancyRepository.deleteById(id);
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        var file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
+    /* При удалении вакансии удаляется связанный с ней файл; */
     @Override
-    public boolean update(Vacancy vacancy) {
-        return vacancyRepository.update(vacancy);
+    public void deleteById(int id) {
+        var fileOptional = findById(id);
+        if (fileOptional.isPresent()) {
+            vacancyRepository.deleteById(id);
+            fileService.deleteById(fileOptional.get().getFileId());
+        }
+    }
+
+    /* При сохранении и обновлении вакансии сохраняется связанный с ней файл */
+    @Override
+    public boolean update(Vacancy vacancy, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return vacancyRepository.update(vacancy);
+        }
+        /* если передан новый не пустой файл, то старый удаляем, а новый сохраняем */
+        var oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        var isUpdated = vacancyRepository.update(vacancy);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
